chore: remove crypto for identity
