+++ conversations/src/group/mod.rs
@@ -1,0 +1,1 @@
+// Group chat module
